+++ lib/input_page.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class InputPage extends StatefulWidget {
+  @override
+  _InputPageState createState() => _InputPageState();
+}
+
+class _InputPageState extends State<InputPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('BMI CALCULATOR'),
+      ),
+      body: Center(
+        child: Text('Body Text'),
+      ),
+    );
+  }
+}
